fix user rating re-rendering
